+++ tests/Codeception/_support/AcceptanceAdminTester.php
@@ -11,6 +11,8 @@ namespace OxidEsales\EshopCommunity\Tests\Codeception;
 
 use Codeception\Actor;
 use OxidEsales\Codeception\Admin\AdminLoginPage;
+use Codeception\Util\Fixtures;
+use OxidEsales\Codeception\Admin\AdminPanel;
 
 /**
  * Inherited Methods
@@ -37,5 +39,12 @@ class AcceptanceAdminTester extends Actor
         $adminLogin = new AdminLoginPage($I);
         $I->amOnPage($adminLogin->URL);
         return $adminLogin;
+    }
+
+    public function loginAdmin(): AdminPanel
+    {
+        $adminPage = $this->openAdmin();
+        $admin = Fixtures::get('adminUser');
+        return $adminPage->login($admin['userLoginName'], $admin['userPassword']);
     }
 }