+++ tests/Codeception/acceptanceAdmin/ModuleNamespaceCest.php
@@ -1,0 +1,91 @@
+<?php
+
+/**
+ * Copyright © OXID eSales AG. All rights reserved.
+ * See LICENSE file for license details.
+ */
+
+declare(strict_types=1);
+
+namespace OxidEsales\EshopCommunity\Tests\CodeceptionAdmin;
+
+use Codeception\Util\Fixtures;
+use OxidEsales\EshopCommunity\Tests\Codeception\AcceptanceAdminTester;
+
+final class ModuleNamespaceCest
+{
+    private $testModule1Path = __DIR__ . '/fixtures/modules/WithNamespaceAndMetadataV2';
+    private $testModule1Id = 'EshopAcceptanceTestModuleNine';
+    private $testModule2Path = __DIR__ . '/fixtures/modules/without_own_module_namespace';
+    private $testModule2Id = 'without_own_module_namespace';
+
+    /** @param AcceptanceAdminTester $I */
+    public function _before(AcceptanceAdminTester $I)
+    {
+        $I->installModule($this->testModule1Path);
+        $I->activateModule($this->testModule1Id);
+
+        $I->installModule($this->testModule2Path);
+        $I->activateModule($this->testModule2Id);
+    }
+
+    /** @param AcceptanceAdminTester $I */
+    public function _after(AcceptanceAdminTester $I)
+    {
+        $I->uninstallModule($this->testModule1Path, $this->testModule1Id);
+        $I->uninstallModule($this->testModule2Path, $this->testModule2Id);
+    }
+        
+    /** @param AcceptanceAdminTester $I */
+    public function testPhysicallyDeleteNamespacedModuleWithoutDeactivation(AcceptanceAdminTester $I): void
+    {
+        $I->wantToTest('Physically remove an activated module from shop without deactivating it.');
+        
+        exec('rm ' . $I->getShopModulePath($this->testModule1Path) . ' -R');
+
+        $this->checkFrontend($I);
+    }
+
+    /** @param AcceptanceAdminTester $I */
+    public function testPhysicallyDeleteNamespacedModuleWithDeactivationAsOnlyModule(AcceptanceAdminTester $I): void
+    {
+        $I->uninstallModule($this->testModule2Path, $this->testModule2Id);
+        $I->wantToTest('Activate then deactivate only module and then physically remove it from shop.');
+        
+        $I->activateModule($this->testModule1Id);
+        $I->deactivateModule($this->testModule1Id);
+        exec('rm ' . $I->getShopModulePath($this->testModule1Path) . ' -R');
+
+        $this->checkFrontend($I);
+
+        $I->installModule($this->testModule2Path);
+    }
+
+    /** @param AcceptanceAdminTester $I */
+    public function testPhysicallyDeleteNamespacedModuleWithDeactivation(AcceptanceAdminTester $I): void
+    {
+        $I->wantToTest('Activate then deactivate a namespace module and then physically remove it from shop.');
+        
+        $I->activateModule($this->testModule1Id);
+        $I->deactivateModule($this->testModule1Id);
+        exec('rm ' . $I->getShopModulePath($this->testModule1Path) . ' -R');
+
+        $this->checkFrontend($I);
+    }
+
+    protected function checkFrontend(AcceptanceAdminTester $I)
+    {
+        $adminPanel = $I->loginAdmin();
+        $moduleList = $adminPanel->openModules();
+
+        $I->see('Problematic Files');
+        $I->see($this->testModule1Id . '/metadata.php');
+
+        $I->click(['name' => 'yesButton']);
+
+        $moduleList->selectModule('Test module #9 - namespaced (EshopAcceptanceTestModuleNine)');
+
+        $I->seeElement('#module_activate');
+        $I->dontSeeElement('#module_deactivate');
+    }
+}